+++ tests/setup_nodes.py
@@ -63,20 +63,6 @@ async def setup_full_node_simulator(db_name, port, introducer_port=None, dic={})
     if db_path.exists():
         db_path.unlink()
 
-    connection = await aiosqlite.connect(db_path)
-    store_1 = await FullNodeStore.create(connection)
-    await store_1._clear_database()
-    unspent_store_1 = await CoinStore.create(connection)
-    await unspent_store_1._clear_database()
-    mempool_1 = MempoolManager(unspent_store_1, test_constants_copy)
-
-    b_1: Blockchain = await Blockchain.create(
-        unspent_store_1, store_1, test_constants_copy
-    )
-    await mempool_1.new_tips(await b_1.get_full_tips())
-
-    await store_1.add_block(FullBlock.from_bytes(test_constants_copy["GENESIS_BLOCK"]))
-
     net_config = load_config(root_path, "config.yaml")
     ping_interval = net_config.get("ping_interval")
     network_id = net_config.get("network_id")
@@ -87,12 +73,8 @@ async def setup_full_node_simulator(db_name, port, introducer_port=None, dic={})
     if introducer_port is not None:
         config["introducer_peer"]["host"] = "127.0.0.1"
         config["introducer_peer"]["port"] = introducer_port
-    full_node_1 = FullNodeSimulator(
-        store_1,
-        b_1,
+    full_node_1 = await FullNodeSimulator.create(
         config,
-        mempool_1,
-        unspent_store_1,
         f"full_node_{port}",
         test_constants_copy,
     )
@@ -114,10 +96,9 @@ async def setup_full_node_simulator(db_name, port, introducer_port=None, dic={})
     yield (full_node_1, server_1)
 
     # TEARDOWN
-    full_node_1._shutdown()
+    await full_node_1._shutdown()
     server_1.close_all()
     await server_1.await_closed()
-    await connection.close()
     db_path.unlink()
 
 
@@ -127,35 +108,19 @@ async def setup_full_node(db_name, port, introducer_port=None, dic={}):
     for k in dic.keys():
         test_constants_copy[k] = dic[k]
 
-    db_path = Path(db_name)
-    connection = await aiosqlite.connect(db_path)
-    store_1 = await FullNodeStore.create(connection)
-    await store_1._clear_database()
-    unspent_store_1 = await CoinStore.create(connection)
-    await unspent_store_1._clear_database()
-    mempool_1 = MempoolManager(unspent_store_1, test_constants_copy)
-
-    b_1: Blockchain = await Blockchain.create(
-        unspent_store_1, store_1, test_constants_copy
-    )
-    await mempool_1.new_tips(await b_1.get_full_tips())
-
-    await store_1.add_block(FullBlock.from_bytes(test_constants_copy["GENESIS_BLOCK"]))
+    Path(db_name).unlink()
 
     net_config = load_config(root_path, "config.yaml")
     ping_interval = net_config.get("ping_interval")
     network_id = net_config.get("network_id")
 
     config = load_config(root_path, "config.yaml", "full_node")
+    config["database_path"] = db_name
     if introducer_port is not None:
         config["introducer_peer"]["host"] = "127.0.0.1"
         config["introducer_peer"]["port"] = introducer_port
-    full_node_1 = FullNode(
-        store_1,
-        b_1,
+    full_node_1 = await FullNode.create(
         config,
-        mempool_1,
-        unspent_store_1,
         f"full_node_{port}",
         test_constants_copy,
     )
